Add Vite config, Vitest configs for CI
The testing still isn't ready enough to turn the GitHub actions back on,
but this is a step in that direction by allowing local testing.

Note that it contains changes to run tests in GitHub Actions both with
jsdom and in the browser (Chrome in CI). We'll see how well this works
when we turn tests back on.