Add first test, import modules on load event
This first test closely resembles the original test from
mbland/tomcat-servlet-testing-example. In fact, it _still_ looks a lot
like it, except that strcalc/src/main/frontend/main.test.js uses a Page
Object. Once this module is published, both that test and this specific
new test will continue to closely resemble one another.

One update to the original JsdomPageOpener that this test required was
to register the module importer as a listener for the "load" event, not
"DOMContentLoaded". I discovered that the tomcat-servlet-testing-example
original loaded a page containing a stylesheet, which caused jsdom not
to have fired "DOMContentLoaded", and document.readyState would be
"loading". The index.html in this test doesn't load any other resources,
so "DOMContentLoaded" fires and document.readyState is "interactive"
before JSDOM.fromFile() even returns.

The solution was to register the importer listener for "load" instead.
This event would already fire before the standalone importModules()
returned, as explained in the long comment updated in this commit.

With this first test now in place, I'm also reenabling GitHub Actions.