Add JsdomPageOpener test
This covers the error handling parts of JsdomPageOpener, and actually
caught a subtle bug in #importModules.

Because the returned Promise wrapped an event handler, the Error thrown
from the standalone importModules() wasn't caught and rejected as
expected.

The fix was to add a `reject` argument to the Promise function, wrap
importModules() in a try/catch, and call `reject` with the caught error.